finishing user authorization

+++ app/Http/Controllers/ListingController.php
@@ -38,8 +38,10 @@ class ListingController extends Controller
         if ($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
-        return redirect('/listings')->with('message', 'List created successfully!');
+        return redirect('/')->with('message', 'List created successfully!');
 
     }
 
@@ -62,7 +64,10 @@ class ListingController extends Controller
     //save updated form
     public function update(Request $request, Listing $listing)
     {
-
+        //Make sure logged-in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -84,7 +89,17 @@ class ListingController extends Controller
     //Delete list
     public function destroy(Listing $listing)
     {
+        //Make sure logged-in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         $listing->delete();
-        return redirect('/listings')->with('message', 'Listing deleted!!');
+        return redirect('/')->with('message', 'Listing deleted!!');
+    }
+
+    //Manage Listings
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
