+++ routes/web.php
@@ -17,29 +17,30 @@ use Illuminate\Support\Facades\Route;
 */
 //==============={Listings Routes}===============
 //All listings
-Route::get('/listings', [ListingController::class, 'index']);
+Route::get('/', [ListingController::class, 'index']);
 //Show create form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 //Store form
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 //Show edit form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 //Save edited form (update the form)[
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 //Delete listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 //Show single list
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
-
+//Manage Listings
+Route::get('/manage/listings',[ListingController::class,'manage'])->middleware('auth');
 
 //==============={User Routes}===============
 //Show register form
-Route::get('/register',[UserController::class,'create']);
+Route::get('/register',[UserController::class,'create'])->middleware('guest');
 //Create new user
 Route::post('/users',[UserController::class,'store']);
 //Log user out
-Route::post('/logout',[UserController::class,'logout']);
+Route::post('/logout',[UserController::class,'logout'])->middleware('auth');
 //Show login form
-Route::get('/login',[UserController::class,'login']);
+Route::get('/login',[UserController::class,'login'])->name('login')->middleware('guest');
 //Log user in
 Route::post('/users/authenticate',[UserController::class,'authenticate']);
